Refactor page.tsx into focused components for better maintainability
Broke down the large page.tsx file into well-scoped, reusable components:
- chat.tsx: Chat interface functionality
- install-prompt.tsx: PWA installation prompts
- meal-planner.tsx: Weekly meal planning features
- push-notification-manager.tsx: Push notification handling

Added supporting utility files in lib/ for shared types and utilities.
This improves code organization, reusability, and maintainability.
