+++ apps/family-app/lib/types.ts
@@ -1,0 +1,4 @@
+export type Meal = {
+	day: string;
+	meal: string;
+};
